requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
pydantic[email]
python-multipart
bcrypt==4.0.1
requests
firebase-admin